python-modules2/requirements.txt
paho-mqtt
python-dateutil
requests
-e libs/docker_secrets